1. Add application name to controller - AD Group 2. Remove SPORelative URL from config 3. Add domain to securitygroup in config 4. Remove unnecessary parameters from config
