+++ SPOWebService/DDMS.WebService.DDMSOperations/Web.config
@@ -31,14 +31,12 @@
     <add key="ClientValidationEnabled" value="true" />
     <add key="UnobtrusiveJavaScriptEnabled" value="true" />
     <add key="SPOSiteURL" value="https://infyakash.sharepoint.com/sites/ddmspoc" />
-    <add key="SPOFolder" value="DDMSData" />
-    <add key="LogDirectory" value="D:\" />
-    <add key="LogPath" value="Log" />
-    <add key="LogLevel" value="DEBUG" />
-    <add key="MaximumFileSizeAllowed" value="31457280" />
-    <add key="SPORelativeURL" value="/sites/ddmspoc" />
-    <add key="Application" value="DDMS"/>
-    <add key="DDMSSecurityGroup" value ="Internet_users_BLR"/>
+    <add key="SPOFolder" value="DDMSData" />    
+    <add key="MaximumFileSizeAllowed" value="31457280" />    
+    
+    <!--Application specific AD Group -->
+    <add key="DDMSSecurityGroup" value ="itlinfosys\Internet_users_BLR"/>
+    
     <add key="SPOUserName" value="PNJkARfnwiaBGDGuXxXZMVo/hZ8MokpCH0QF1imJB+Pqvx6w/Bh8CMOikhpT4Vvv" />
     <add key="SPOUserNameKey" value="8OZYc++sDXGHFCZ8BP8pHP+uAOZ6P9mqSFBMHmWAccU=" />
     <add key="SPOUserNameIv" value="o3bfHTRCB5iaduwRaLN2CA==" />
